requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.11.0
pillow>=8.3.0
pyyaml>=5.4.1
matplotlib>=3.4.3
seaborn>=0.11.2
opencv-python>=4.5.3
tensorboard>=2.6.0
tqdm>=4.62.3